Add a persistent struct to query versions for
This adds a `Rustc` struct which can be used to ensure that the rustc
command line is executed only once and then is a little bit more
ergonomic for making multiple repeated queries to over time. The
intention here is to make it easy to ask multiple "is the compiler at
least this version?" questions without having to run rustc more than
once and still using the built-in APIs here.